Remove sh from busybox.links.nosuid
I hope that this fix busybox update.
It looks like the trick with sh copying in tmp does not work and after update-alternatives remove sh, it stops working.
Link to sh already exists in the package, if remove it from the busybox.links.nosuid, it is not used in update-alternatives.